+++ R/buildSimList.R
@@ -21,27 +21,27 @@
 #' or terra objects.
 #' 
 #' @param roads roads input
-#' @param cost cost input
+#' @param weightRaster weightRaster input
 #' @param roadMethod method of road projection
 #' @param landings landings input
-#' @param roadsInCost Whether the roads have already been burned into cost
+#' @param roadsInWeight Whether the roads have already been burned into the `weightRaster`.
 #' @param sim A sim list to update rather than building new.
 #' @noRd
 
-buildSimList <- function(roads, cost, roadMethod, landings, roadsInCost, 
+buildSimList <- function(roads, weightRaster, roadMethod, landings, roadsInWeight, 
                          sim = NULL){
-  if(!is(cost, "SpatRaster")){
-    if(is(cost, "RasterLayer")){
-      cost <- terra::rast(cost)
+  if(!is(weightRaster, "SpatRaster")){
+    if(is(weightRaster, "RasterLayer")){
+      weightRaster <- terra::rast(weightRaster)
     } else {
-      stop("cost must be provided as a SpatRaster or RasterLayer", call. = FALSE)
+      stop("weightRaster must be provided as a SpatRaster or RasterLayer", call. = FALSE)
     }
   } 
-  if(terra::minmax(cost)[1] == 0 || 0 %in% terra::unique(cost)[[1]]){
-    message("0s detected in cost raster, these will be considered as existing roads")
-  } else if(roadsInCost){
-    warning("No 0s detected in cost raster. If existing roads have not been ",
-            "included in the cost raster set roadsInCost = FALSE to have them ",
+  if(terra::minmax(weightRaster)[1] == 0 || 0 %in% terra::unique(weightRaster)[[1]]){
+    message("0s detected in weightRaster raster, these will be considered as existing roads")
+  } else if(roadsInWeight){
+    warning("No 0s detected in weightRaster. If existing roads have not been ",
+            "included in the weightRaster set roadsInWeight = FALSE to have them ",
             "burned in", call. = FALSE)
   }
   
@@ -50,10 +50,10 @@ buildSimList <- function(roads, cost, roadMethod, landings, roadsInCost,
   }
   
   # Burn roads into raster if not already for raster roads before converting 
-  if(!roadsInCost && is(roads, "SpatRaster")){
-    message("Burning in roads to cost raster from road raster")
-    cost <- cost * (roads == 0)
-    roadsInCost <- TRUE
+  if(!roadsInWeight && is(roads, "SpatRaster")){
+    message("Burning in roads to weightRaster raster from road raster")
+    weightRaster <- weightRaster * (roads == 0)
+    roadsInWeight <- TRUE
   }
   
   if(!(is(roads, "sf") || is(roads, "sfc"))){
@@ -121,9 +121,9 @@ buildSimList <- function(roads, cost, roadMethod, landings, roadsInCost,
         ) %>%
         sf::st_cast("POINT") %>% 
         sf::st_set_agr("constant") %>% 
-        sf::st_set_crs(sf::st_crs(cost))
+        sf::st_set_crs(sf::st_crs(weightRaster))
       
-      message("CRS of landings supplied as a matrix is assumed to match the cost")
+      message("CRS of landings supplied as a matrix is assumed to match the weightRaster")
     }else {
       stop("landings must be either SpatRaster, RasterLayer, sf object, SpatialPoints*, ",
            "or SpatialPolygons*",
@@ -144,19 +144,19 @@ buildSimList <- function(roads, cost, roadMethod, landings, roadsInCost,
   
   # check crs error if different
   if(!all(sf::st_crs(roads) == sf::st_crs(landings),
-          sf::st_crs(roads) == sf::st_crs(cost))){
-    stop("the crs of roads, landings and cost must match", call. = FALSE)
+          sf::st_crs(roads) == sf::st_crs(weightRaster))){
+    stop("the crs of roads, landings and weightRaster must match", call. = FALSE)
   }
   
-  # burn in roads to have 0 cost 
-  if(!roadsInCost){
-    message("Burning in roads to cost raster from sf")
+  # burn in roads to have 0 weight 
+  if(!roadsInWeight){
+    message("Burning in roads to weightRaster from sf")
     
-    cost <- burnRoadsInCost(roads, cost)
+    weightRaster <- burnRoadsInWeight(roads, weightRaster)
     
   }
   
-  # crop landings and roads to bbox of cost raster
+  # crop landings and roads to bbox of weightRaster raster
   nrland <- nrow(landings)
   nrroads <- nrow(roads)
   
@@ -164,29 +164,29 @@ buildSimList <- function(roads, cost, roadMethod, landings, roadsInCost,
     stop("nrow(roads) is 0. Please supply at least one existing road", call. = FALSE)
   }
   
-  ext <- sf::st_bbox(cost) %>% as.numeric() %>% 
+  ext <- sf::st_bbox(weightRaster) %>% as.numeric() %>% 
     `names<-`(c("xmin", "ymin", "xmax", "ymax"))
   landings <- sf::st_crop(sf::st_set_agr(landings, "constant"), ext) %>% sf::st_set_agr("constant")
   roads <- sf::st_crop(sf::st_set_agr(roads, "constant"), ext) %>% sf::st_set_agr("constant")
   
   if(nrland != nrow(landings)){
-    stop(nrland - nrow(landings), " landings are outside the cost surface. ",
-         "The cost surface must cover the extent of landings", call. = FALSE)
+    stop(nrland - nrow(landings), " landings are outside the weightRaster. ",
+         "The weightRaster must cover the extent of landings", call. = FALSE)
   }
   
   if(nrroads != nrow(roads)){
-    stop(nrroads - nrow(roads), " roads are outside the cost surface. ",
-         "The cost surface must cover the extent of roads", call. = FALSE)
+    stop(nrroads - nrow(roads), " roads are outside the weightRaster. ",
+         "The weightRaster must cover the extent of roads", call. = FALSE)
   }
 
   if(is.null(sim)){
-    sim <- rlang::env(roads = roads, costSurface = cost, 
+    sim <- rlang::env(roads = roads, weightRaster = weightRaster, 
                 roadMethod = roadMethod, 
                 landings = landings)
   } else {
     sim$roads <- roads
     sim$landings <- landings
-    sim$costSurface <- cost
+    sim$weightRaster <- weightRaster
     sim$roadMethod <- roadMethod
   }
   return(sim)
@@ -195,37 +195,37 @@ buildSimList <- function(roads, cost, roadMethod, landings, roadsInCost,
 
 
 
-#' Burn roads in to cost
+#' Burn roads in to weightRaster
 #'
-#' Use sf roads object to convert cost to 0 where roads already exist. This is
+#' Use sf roads object to convert weightRaster to 0 where roads already exist. This is
 #' an internal function and does not contain many checks. Use at own risk.
 #'
 #' @param roads sf object with road lines
-#' @param cost SpatRaster with cost of road development
+#' @param weightRaster SpatRaster with weights
 #'
-#' @return SpatRaster of cost with 0 for roads.
+#' @return a SpatRaster with 0 for roads.
 #' 
 #' @noRd
-burnRoadsInCost <- function(roads, cost){
+burnRoadsInWeight <- function(roads, weightRaster){
   # The crs is checked above but stars requires that they be identical
   if(!is.na(sf::st_crs(roads))){
-    roads <- sf::st_transform(roads, sf::st_crs(cost))
+    roads <- sf::st_transform(roads, sf::st_crs(weightRaster))
   }
   
   if(any(grepl("POINT", sf::st_geometry_type(roads, by_geometry = TRUE))) &&
      any(grepl("LINESTRING", sf::st_geometry_type(roads, by_geometry = TRUE)))){
     geom_types <- c("POINT", "LINESTRING")
     
-    rasts <- lapply(geom_types, function(x, rds, cst){
+    rasts <- lapply(geom_types, function(x, rds, wt){
       geom_roads <- sf::st_collection_extract(rds, type = x)
-      geom_rast <- terra::rasterize(terra::vect(geom_roads), cst,
+      geom_rast <- terra::rasterize(terra::vect(geom_roads), wt,
                                     background = 0) > 0
-    }, rds = roads, cst = cost)
+    }, rds = roads, wt = weightRaster)
     
     roadsRast <- !(rasts[[1]]|rasts[[2]])
   } else {
-    roadsRast <- terra::rasterize(terra::vect(roads), cost, background = 0) == 0 
+    roadsRast <- terra::rasterize(terra::vect(roads), weightRaster, background = 0) == 0 
   }
   
-  cost <- cost * roadsRast
+  weightRaster <- weightRaster * roadsRast
 }