+++ R/plotRoads.R
@@ -21,7 +21,7 @@
 #' @param sim sim list result from `projectRoads`
 #' @param mainTitle A title for the plot
 #' @param subTitle A sub title for the plot, by default the `roadMethod` is used
-#' @param ... Other arguments passed to raster plot call for the `costSurface`
+#' @param ... Other arguments passed to raster plot call for the `weightRaster`
 #' 
 #' @return Creates a plot using base graphics
 #' 
@@ -38,7 +38,7 @@
 plotRoads <- function(sim, mainTitle, 
                       subTitle = paste0("Method: ", sim$roadMethod), 
                       ...){
-  terra::plot(sim$costSurface, main = mainTitle, 
+  terra::plot(sim$weightRaster, main = mainTitle, 
                sub = subTitle, ...)
   if(is(sim$roads, "SpatRaster")){
     terra::plot(terra::subst(sim$roads, from = c(0,1), to = c(NA, 1)),