+++ R/weightFunctions.R
@@ -7,6 +7,7 @@
 #'   * Negative values are costs for crossing streams or other barriers that are
 #'   crossable but expensive. Edges that link to barrier penalty (negative value)
 #'   nodes are assigned the largest barrier penalty weight.
+#'   * Zero values are assumed to be existing roads.
 #'   * All other values are interpreted as elevation in the units of the raster
 #'    map (so that a difference between two cells equal to the map resolution can be
 #'     interpreted as 100% grade)
@@ -21,7 +22,7 @@
 #' @param baseCost Construction cost of 0% grade road.
 #' @param limit Maximum grade (%) on which roads can be built.
 #' @param penalty Cost increase associated with each additional % increase in road grade.
-#' @param limitCost Value assigned to edges that exceed the grade limit. Try setting to a high (not NA) value if encountering problems with disconnected graphs.
+#' @param limitWeight Value assigned to edges that exceed the grade limit. Try setting to a high (not NA) value if encountering problems with disconnected graphs.
 #'
 #' @export
 #'
@@ -30,14 +31,14 @@
 #' slopePenaltyFn(0.5,0.65)
 #' # grade > 20% so NA
 #' slopePenaltyFn(0.5,0.75)
-slopePenaltyFn<-function(x1,x2,resolution=1, baseCost = 16178,limit=20,penalty=504,limitCost=NA){
+slopePenaltyFn<-function(x1,x2,resolution=1, baseCost = 16178,limit=20,penalty=504,limitWeight=NA){
   #If one of the nodes is a road or barrier ignore grade penalty
   cond = pmin(x1,x2)>0
   cond[is.na(cond)]=F
   grade = 100*abs(x1-x2)*cond/resolution #percent slope, if both nodes have elevation values.
   #stop(print(max(grade)))
   slp = baseCost+grade*penalty
-  slp[grade>limit]=limitCost
+  slp[grade>limit]=limitWeight
 
   slp[!cond] = abs(pmin(x1,x2))[!cond] # if both 0 this is an existing road link. Otherwise it is a barrier.
   return(slp)